[go] update
